Read flag from env
I tried to make clap parse this but by default clap will treat any
value, including "", "0", and "false" as true. Every way I tried to
change this would introduce some other problem. For instance changing
the action to set would make it show up as a positional argument.

The values accepted are limited to "0" and "1" because this is
sufficient to communicate all options, and adding more accepted values
is easier than removing them.